requirements.txt
langchain
langchain-community
streamlit
python-docx
langchain-google-genai
python-dotenv
